collatz_sequence
The Collatz sequence.
Start with any integer.
If it's even, divide by two.
If it's odd, multiply it by three and add one.
Repeat until the number is 1.

Currently this is not proven, but no one has been able to disprove it. There is an international prize for proving this.. problem. :)